// src/app/shared/types/api-types/bidt-stock-mvt-type-dto.interface.ts
// //////////////////////////////////////////////////////////////////////////
// DO NOT EDIT THIS FILE
//
// It is automatically generated by generate-dtos.js
// //////////////////////////////////////////////////////////////////////////

import { BidtStockMvtDTO } from './bidt-stock-mvt-dto.interface';
import { BidtStockMvtTypeIdDTO } from './bidt-stock-mvt-type-id-dto.interface';
import { BidtStockMvtTypeOperationDTO } from './bidt-stock-mvt-type-operation-dto.interface';

export interface BidtStockMvtTypeDTO extends BidtStockMvtTypeIdDTO {
  bidtStockMvts?: BidtStockMvtDTO[];
  id?: number;
  bidtStockMvtTypeOperations?: BidtStockMvtTypeOperationDTO[];
  issueAreaType?: string; // 50
  receiptAreaType?: string; // 50
  smtName?: string; // 100
  smtStatuses?: string; // 300
  smtWay?: string; // 50
}