// src/app/shared/types/api-types/bidt-stock-mvt-type-operation-dto.interface.ts
// //////////////////////////////////////////////////////////////////////////
// DO NOT EDIT THIS FILE
//
// It is automatically generated by generate-dtos.js
// //////////////////////////////////////////////////////////////////////////

import { BidtStockMvtTypeOperationIdDTO } from './bidt-stock-mvt-type-operation-id-dto.interface';

export interface BidtStockMvtTypeOperationDTO extends BidtStockMvtTypeOperationIdDTO {
  bidtStockMvtType: string;

  smtoKey?: string; // 50
  smtoSequenceNumber?: string;
  taskTypeCode?: string; // 50
  pnTypes: string[];
  docCategories: string[];
}